requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.3.0
transformers>=4.30.0
kafka-python>=2.0.2
fastapi>=0.100.0
uvicorn>=0.23.0
redis>=4.6.0
shap>=0.42.0
captum>=0.6.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
tqdm>=4.65.0
pyyaml>=6.0
joblib>=1.3.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
networkx>=3.1
dgl>=1.1.0
optuna>=3.0.0
